+++ example/lib/main.dart
@@ -1,11 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_window_close/flutter_window_close.dart';
-import 'package:livekit_client/livekit_client.dart';
 import 'package:livekit_example/theme.dart';
 import 'package:logging/logging.dart';
 import 'package:intl/intl.dart';
 import 'pages/connect.dart';
-import 'utils.dart';
 
 void main() async {
   final format = DateFormat('HH:mm:ss');
@@ -16,13 +13,12 @@ void main() async {
   });
 
   WidgetsFlutterBinding.ensureInitialized();
-
-  if (lkPlatformIsDesktop()) {
+  /*if (lkPlatformIsDesktop()) {
     await FlutterWindowClose.setWindowShouldCloseHandler(() async {
       await onWindowShouldClose?.call();
       return true;
     });
-  }
+  }*/
 
   /// for livestreaming app, you can initialize the bypassVoiceProcessing = true
   /// here to get better audio quality
@@ -30,7 +26,6 @@ void main() async {
   /// await LiveKitClient.initialize(
   ///  bypassVoiceProcessing: lkPlatformIsMobile(),
   /// );
-
   runApp(const LiveKitExampleApp());
 }