+++ lib/src/proto/livekit_models.pbenum.dart
@@ -213,6 +213,12 @@ class DisconnectReason extends $pb.ProtobufEnum {
       DisconnectReason._(9, _omitEnumNames ? '' : 'SIGNAL_CLOSE');
   static const DisconnectReason ROOM_CLOSED =
       DisconnectReason._(10, _omitEnumNames ? '' : 'ROOM_CLOSED');
+  static const DisconnectReason USER_UNAVAILABLE =
+      DisconnectReason._(11, _omitEnumNames ? '' : 'USER_UNAVAILABLE');
+  static const DisconnectReason USER_REJECTED =
+      DisconnectReason._(12, _omitEnumNames ? '' : 'USER_REJECTED');
+  static const DisconnectReason SIP_TRUNK_FAILURE =
+      DisconnectReason._(13, _omitEnumNames ? '' : 'SIP_TRUNK_FAILURE');
 
   static const $core.List<DisconnectReason> values = <DisconnectReason>[
     UNKNOWN_REASON,
@@ -226,6 +232,9 @@ class DisconnectReason extends $pb.ProtobufEnum {
     MIGRATION,
     SIGNAL_CLOSE,
     ROOM_CLOSED,
+    USER_UNAVAILABLE,
+    USER_REJECTED,
+    SIP_TRUNK_FAILURE,
   ];
 
   static final $core.Map<$core.int, DisconnectReason> _byValue =
@@ -446,6 +455,10 @@ class ClientInfo_SDK extends $pb.ProtobufEnum {
       ClientInfo_SDK._(9, _omitEnumNames ? '' : 'PYTHON');
   static const ClientInfo_SDK CPP =
       ClientInfo_SDK._(10, _omitEnumNames ? '' : 'CPP');
+  static const ClientInfo_SDK UNITY_WEB =
+      ClientInfo_SDK._(11, _omitEnumNames ? '' : 'UNITY_WEB');
+  static const ClientInfo_SDK NODE =
+      ClientInfo_SDK._(12, _omitEnumNames ? '' : 'NODE');
 
   static const $core.List<ClientInfo_SDK> values = <ClientInfo_SDK>[
     UNKNOWN,
@@ -459,6 +472,8 @@ class ClientInfo_SDK extends $pb.ProtobufEnum {
     RUST,
     PYTHON,
     CPP,
+    UNITY_WEB,
+    NODE,
   ];
 
   static final $core.Map<$core.int, ClientInfo_SDK> _byValue =
@@ -466,6 +481,32 @@ class ClientInfo_SDK extends $pb.ProtobufEnum {
   static ClientInfo_SDK? valueOf($core.int value) => _byValue[value];
 
   const ClientInfo_SDK._($core.int v, $core.String n) : super(v, n);
+}
+
+/// enum for operation types (specific to TextHeader)
+class DataStream_OperationType extends $pb.ProtobufEnum {
+  static const DataStream_OperationType CREATE =
+      DataStream_OperationType._(0, _omitEnumNames ? '' : 'CREATE');
+  static const DataStream_OperationType UPDATE =
+      DataStream_OperationType._(1, _omitEnumNames ? '' : 'UPDATE');
+  static const DataStream_OperationType DELETE =
+      DataStream_OperationType._(2, _omitEnumNames ? '' : 'DELETE');
+  static const DataStream_OperationType REACTION =
+      DataStream_OperationType._(3, _omitEnumNames ? '' : 'REACTION');
+
+  static const $core.List<DataStream_OperationType> values =
+      <DataStream_OperationType>[
+    CREATE,
+    UPDATE,
+    DELETE,
+    REACTION,
+  ];
+
+  static final $core.Map<$core.int, DataStream_OperationType> _byValue =
+      $pb.ProtobufEnum.initByValue(values);
+  static DataStream_OperationType? valueOf($core.int value) => _byValue[value];
+
+  const DataStream_OperationType._($core.int v, $core.String n) : super(v, n);
 }
 
 const _omitEnumNames = $core.bool.fromEnvironment('protobuf.omit_enum_names');