+++ lib/src/support/native_audio.dart
@@ -86,15 +86,50 @@ class NativeAudioConfiguration {
   final AppleAudioCategory? appleAudioCategory;
   final Set<AppleAudioCategoryOption>? appleAudioCategoryOptions;
   final AppleAudioMode? appleAudioMode;
+  final bool? preferSpeakerOutput;
 
-  NativeAudioConfiguration({
-    // for iOS / Mac
-    this.appleAudioCategory,
-    this.appleAudioCategoryOptions,
-    this.appleAudioMode,
-    // Android options
-    // ...
-  });
+  static final soloAmbient = NativeAudioConfiguration(
+    appleAudioCategory: AppleAudioCategory.soloAmbient,
+    appleAudioCategoryOptions: {},
+    appleAudioMode: AppleAudioMode.default_,
+  );
+
+  static final playback = NativeAudioConfiguration(
+    appleAudioCategory: AppleAudioCategory.playback,
+    appleAudioCategoryOptions: {AppleAudioCategoryOption.mixWithOthers},
+    appleAudioMode: AppleAudioMode.spokenAudio,
+  );
+
+  static final playAndRecordSpeaker = NativeAudioConfiguration(
+    appleAudioCategory: AppleAudioCategory.playAndRecord,
+    appleAudioCategoryOptions: {
+      AppleAudioCategoryOption.allowBluetooth,
+      AppleAudioCategoryOption.allowBluetoothA2DP,
+      AppleAudioCategoryOption.allowAirPlay,
+    },
+    appleAudioMode: AppleAudioMode.videoChat,
+  );
+
+  static final playAndRecordReceiver = NativeAudioConfiguration(
+    appleAudioCategory: AppleAudioCategory.playAndRecord,
+    appleAudioCategoryOptions: {
+      AppleAudioCategoryOption.allowBluetooth,
+      AppleAudioCategoryOption.allowBluetoothA2DP,
+      AppleAudioCategoryOption.allowAirPlay,
+    },
+    appleAudioMode: AppleAudioMode.voiceChat,
+  );
+
+  NativeAudioConfiguration(
+      {
+      // for iOS / Mac
+      this.appleAudioCategory,
+      this.appleAudioCategoryOptions,
+      this.appleAudioMode,
+      this.preferSpeakerOutput
+      // Android options
+      // ...
+      });
 
   Map<String, dynamic> toMap() => <String, dynamic>{
         if (appleAudioCategory != null)
@@ -104,17 +139,21 @@ class NativeAudioConfiguration {
               appleAudioCategoryOptions!.map((e) => e.toStringValue()).toList(),
         if (appleAudioMode != null)
           'appleAudioMode': appleAudioMode!.toStringValue(),
+        if (preferSpeakerOutput != null)
+          'preferSpeakerOutput': preferSpeakerOutput,
       };
 
   NativeAudioConfiguration copyWith({
     AppleAudioCategory? appleAudioCategory,
     Set<AppleAudioCategoryOption>? appleAudioCategoryOptions,
     AppleAudioMode? appleAudioMode,
+    bool? preferSpeakerOutput,
   }) =>
       NativeAudioConfiguration(
         appleAudioCategory: appleAudioCategory ?? this.appleAudioCategory,
         appleAudioCategoryOptions:
             appleAudioCategoryOptions ?? this.appleAudioCategoryOptions,
         appleAudioMode: appleAudioMode ?? this.appleAudioMode,
+        preferSpeakerOutput: preferSpeakerOutput ?? this.preferSpeakerOutput,
       );
 }