+++ lib/src/track/web/_audio_context.dart
@@ -12,13 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'dart:js_util' as jsutil;
-
-import 'package:js/js.dart';
-
-@JS()
-external dynamic get undefined;
-
 enum AudioContextState {
   suspended('suspended'),
   running('running'),
@@ -30,26 +23,4 @@ enum AudioContextState {
   static Iterable<AudioContextState> fromValues(Iterable<String> values) =>
       values.map(fromValue);
   const AudioContextState(this.value);
-}
-
-@anonymous
-@JS()
-@staticInterop
-class AudioContextOptions {
-  external factory AudioContextOptions(
-      {dynamic latencyHint, double? sampleRate});
-}
-
-@JS()
-@staticInterop
-class AudioContext {
-  external factory AudioContext._([AudioContextOptions? contextOptions]);
-
-  factory AudioContext([AudioContextOptions? contextOptions]) =>
-      AudioContext._(contextOptions ?? undefined);
-}
-
-extension PropsBaseAudioContext on AudioContext {
-  AudioContextState get state =>
-      AudioContextState.fromValue(jsutil.getProperty(this, 'state'));
 }