+++ shared_swift/LiveKitPlugin.swift
@@ -1,4 +1,4 @@
-// Copyright 2024 LiveKit, Inc.
+// Copyright 2025 LiveKit, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,13 +21,20 @@ import FlutterMacOS
 #else
 import Flutter
 import UIKit
+import Combine
 #endif
 
+@available(iOS 13.0, *)
 public class LiveKitPlugin: NSObject, FlutterPlugin {
 
-    var processers: Dictionary<Track, Visualizer> = [:]
-    
+    var processors: Dictionary<String, Visualizer> = [:]
+    var tracks: Dictionary<String, Track> = [:]
+
     var binaryMessenger: FlutterBinaryMessenger?
+
+    #if os(iOS)
+    var cancellable = Set<AnyCancellable>()
+    #endif
 
     public static func register(with registrar: FlutterPluginRegistrar) {
 
@@ -41,6 +48,14 @@ public class LiveKitPlugin: NSObject, FlutterPlugin {
         let instance = LiveKitPlugin()
         instance.binaryMessenger = messenger
         registrar.addMethodCallDelegate(instance, channel: channel)
+
+        #if os(iOS)
+        BroadcastManager.shared.isBroadcastingPublisher
+            .sink { isBroadcasting in
+                channel.invokeMethod("broadcastStateChanged", arguments: isBroadcasting)
+            }
+            .store(in: &instance.cancellable)
+        #endif
     }
 
     #if !os(macOS)
@@ -90,49 +105,66 @@ public class LiveKitPlugin: NSObject, FlutterPlugin {
         return result
     }
     #endif
-    
+
     public func handleStartAudioVisualizer(args: [String: Any?], result: @escaping FlutterResult) {
         let webrtc = FlutterWebRTCPlugin.sharedSingleton()
-        
+
         let trackId = args["trackId"] as? String
+        let visualizerId = args["visualizerId"] as? String 
         let barCount = args["barCount"] as? Int ?? 7
         let isCentered = args["isCentered"] as? Bool ?? true
-        
-        if let unwrappedTrackId = trackId {
-            
+
+        if visualizerId == nil {
+            result(FlutterError(code: "visualizerId", message: "visualizerId is required", details: nil))
+            return
+        }
+
+        if let unwrappedTrackId = trackId { 
+            let unwrappedVisualizerId = visualizerId!
+
             let localTrack = webrtc?.localTracks![unwrappedTrackId]
             if let audioTrack = localTrack as? LocalAudioTrack {
                 let lkLocalTrack = LKLocalAudioTrack(name: unwrappedTrackId, track: audioTrack);
                 let processor = Visualizer(track: lkLocalTrack,
                                                binaryMessenger: self.binaryMessenger!,
                                                bandCount: barCount,
-                                               isCentered: isCentered)
-                processers[lkLocalTrack] = processor
+                                               isCentered: isCentered,
+                                               visualizerId: unwrappedVisualizerId)    
+                
+                tracks[unwrappedTrackId] = lkLocalTrack
+                processors[unwrappedVisualizerId] = processor
+                
             }
-             
+
             let track = webrtc?.remoteTrack(forId: unwrappedTrackId)
             if let audioTrack = track as? RTCAudioTrack {
                 let lkRemoteTrack = LKRemoteAudioTrack(name: unwrappedTrackId, track: audioTrack);
                 let processor = Visualizer(track: lkRemoteTrack,
                                                binaryMessenger: self.binaryMessenger!,
                                                bandCount: barCount,
-                                               isCentered: isCentered)
-                processers[lkRemoteTrack] = processor
+                                               isCentered: isCentered,
+                                               visualizerId: unwrappedVisualizerId)
+                tracks[unwrappedTrackId] = lkRemoteTrack
+                processors[unwrappedVisualizerId] = processor
             }
         }
-        
-        
+
+
         result(true)
     }
-    
+
     public func handleStopAudioVisualizer(args: [String: Any?], result: @escaping FlutterResult) {
         let trackId = args["trackId"] as? String
+        let visualizerId = args["visualizerId"] as? String
         if let unwrappedTrackId = trackId {
-            for key in processers.keys {
-                if key.mediaTrack.trackId == unwrappedTrackId {
-                    processers.removeValue(forKey: key)
+            for key in tracks.keys {
+                if key == unwrappedTrackId {
+                    tracks.removeValue(forKey: key)
                 }
             }
+        }
+        if let unwrappedVisualizerId = visualizerId {
+            processors.removeValue(forKey: unwrappedVisualizerId)
         }
         result(true)
     }
@@ -198,6 +230,10 @@ public class LiveKitPlugin: NSObject, FlutterPlugin {
             //                      options: configuration.categoryOptions)
             // print("[LiveKit] AVAudioSession Configure success")
 
+            // preferSpeakerOutput
+            if let preferSpeakerOutput = args["preferSpeakerOutput"] as? Bool {
+              try rtcSession.overrideOutputAudioPort(preferSpeakerOutput ? .speaker : .none)
+            }
             result(true)
         } catch let error {
             print("[LiveKit] Configure audio error: ", error)
@@ -223,7 +259,6 @@ public class LiveKitPlugin: NSObject, FlutterPlugin {
     }
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-
         guard let args = call.arguments as? [String: Any?] else {
             print("[LiveKit] arguments must be a dictionary")
             result(FlutterMethodNotImplemented)
@@ -239,6 +274,14 @@ public class LiveKitPlugin: NSObject, FlutterPlugin {
             handleStopAudioVisualizer(args: args, result: result)
         case "osVersionString":
             result(LiveKitPlugin.osVersionString())
+        #if os(iOS)
+        case "broadcastRequestActivation":
+            BroadcastManager.shared.requestActivation()
+            result(true)
+        case "broadcastRequestStop":
+            BroadcastManager.shared.requestStop()
+            result(true)
+        #endif
         default:
             print("[LiveKit] method not found: ", call.method)
             result(FlutterMethodNotImplemented)